requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
charset-normalizer==2.1.1
dj-database-url==1.2.0
Django==4.1.4
idna==3.4
numpy==1.24.0
opencv-python==4.6.0.66
Pillow==9.3.0
psycopg2==2.9.5
requests==2.28.1
sqlparse==0.4.3
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.4.0
urllib3==1.26.13
whitenoise==6.3.0
